+++ examples/basic/app/routes/api/image.ts
@@ -1,24 +1,26 @@
 import type { LoaderFunction } from "remix";
 import {
   imageLoader,
-  MemoryCache,
+  DiskCache,
   fsResolver,
   fetchResolver,
   Resolver,
 } from "remix-image/server";
+import { sharpTransformer } from "../../utils/sharp";
 
-export const fetchImage: Resolver = async (asset, url, options) => {
+export const fetchImage: Resolver = async (asset, url, options, basePath) => {
   if (url.startsWith("/") && (url.length === 1 || url[1] !== "/")) {
-    return fsResolver(asset, url, options);
+    return fsResolver(asset, url, options, basePath);
   } else {
-    return fetchResolver(asset, url, options);
+    return fetchResolver(asset, url, options, basePath);
   }
 };
 
 const config = {
   selfUrl: "http://localhost:3000",
-  cache: new MemoryCache(),
+  cache: new DiskCache(),
   resolver: fetchImage,
+  transformer: sharpTransformer,
 };
 
 export const loader: LoaderFunction = ({ request }) => {