+++ examples/vercel/app/routes/api/image.ts
@@ -1,12 +1,14 @@
+import os from "os";
+import path from "path";
 import type { LoaderFunction } from "remix";
 import {
   imageLoader,
+  DiskCache,
   fsResolver,
   fetchResolver,
   Resolver,
-  MemoryCache,
 } from "remix-image/server";
-import { sharpTransformer } from "../../../sharpTransformer";
+import { sharpTransformer } from "../../utils/sharp";
 
 export const fetchImage: Resolver = async (asset, url, options, basePath) => {
   if (url.startsWith("/") && (url.length === 1 || url[1] !== "/")) {
@@ -16,11 +18,22 @@ export const fetchImage: Resolver = async (asset, url, options, basePath) => {
   }
 };
 
+const vercelUrl = process.env.VERCEL_URL || "";
+const fixedVercelUrl = vercelUrl.startsWith("https")
+  ? vercelUrl
+  : `https://${vercelUrl}`;
+
 const config = {
-  selfUrl: "http://localhost:3000",
-  cache: new MemoryCache(),
-  transformer: sharpTransformer,
+  selfUrl:
+    process.env.NODE_ENV === "development"
+      ? "http://localhost:3000"
+      : fixedVercelUrl,
+  cache: new DiskCache({
+    path: path.join(os.tmpdir(), "img"),
+  }),
   resolver: fetchImage,
+  transformer: sharpTransformer,
+  basePath: process.env.NODE_ENV === "development" ? "public" : "/",
 };
 
 export const loader: LoaderFunction = ({ request }) => {