+++ src/server/caches/diskCache.ts
@@ -1,4 +1,4 @@
-import BaseCache from "hybrid-disk-cache";
+import BaseCache from "@next-boost/hybrid-disk-cache";
 import { CacheConfig, Cache, CacheStatus } from "../../types/cache";
 
 export interface DiskCacheConfig extends CacheConfig {