+++ CHANGELOG.md
@@ -3,6 +3,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `basePath` property to `LoaderConfig` and `Resolver` to allow custom base paths for assets
+
+### Changed
+
+- Replaced `hybrid-disk-cache` package with the updated version `@next-boost/hybrid-disk-cache`
+
+### Documentation
+
+- Added Vercel deployment example
+- Replaced functions in examples with best practices (file cache, sharp)
+
 ## [0.3.10] - 2022-02-17
 
 ### Documentation